Share DigitalTop/ChipyardSystem | Fix small naming compile error
